+++ spec/yaml_helper_spec.rb
@@ -1,74 +1,25 @@
 require File.expand_path('../spec_helper', __FILE__)
 
-def sample_yaml
-  <<-LOCKFILE.strip_heredoc
-      PODS:
-        - BananaLib (1.0):
-          - monkey (< 1.0.9, ~> 1.0.1)
-        - JSONKit (1.4)
-        - monkey (1.0.8)
-
-      DEPENDENCIES:
-        - BananaLib (~> 1.0)
-        - JSONKit (from `path/JSONKit.podspec`)
-
-      EXTERNAL SOURCES:
-        JSONKit:
-          podspec: path/JSONKit.podspec
-
-      SPEC CHECKSUMS:
-        BananaLib: 439d9f683377ecf4a27de43e8cf3bce6be4df97b
-        JSONKit: 92ae5f71b77c8dec0cd8d0744adab79d38560949
-
-      COCOAPODS: 1.0.0
-  LOCKFILE
-end
-
-def yaml_with_merge_conflict
-  <<-LOCKFILE.strip_heredoc
-    PODS:
-      - Kiwi (2.2)
-      - ObjectiveSugar (1.1.1)
-
-    DEPENDENCIES:
-      - Kiwi
-      - ObjectiveSugar (from `../`)
-
-    EXTERNAL SOURCES:
-      ObjectiveSugar:
-        :path: ../
-
-    SPEC CHECKSUMS:
-    <<<<<<< HEAD
-      Kiwi: 05f988748c5136c6daed8dab3563eca929399a72
-      ObjectiveSugar: 7377622e35ec89ce893b05dd0af4bede211b01a4
-    =======
-      Kiwi: db174bba4ee8068b15d7122f1b22fb64b7c1d378
-      ObjectiveSugar: 27c680bb74f0b0415e9e743d5d61d77bc3292d3f
-    >>>>>>> b65623cbf5e105acbc3e2dec48f8024fa82003ce
-
-    COCOAPODS: 0.29.0
-  LOCKFILE
-end
-
-def bad_yaml
-  <<-LOCKFILE.strip_heredoc
-    PODS:
-      - Kiwi (2.2)
-      SOME BAD TEXT
-
-    DEPENDENCIES:
-      - Kiwi
-      - ObjectiveSugar (from `../`)
-
-    COCOAPODS: 0.29.0
-  LOCKFILE
-end
-
-#-----------------------------------------------------------------------------#
-
 module Pod
   describe 'In general' do
+
+    before do
+      @good_podfile_lock      = File.open(
+        File.expand_path('../fixtures/Podfile.lock', __FILE__))
+
+      @bad_yaml_podfile_lock  = File.open(
+        File.expand_path('../fixtures/PodfileWithIncorrectYAML.lock', __FILE__))
+
+      @conflict_podfile_lock  = File.open(
+        File.expand_path('../fixtures/PodfileWithMergeConflicts.lock', __FILE__))
+
+    end
+
+    after do
+      @good_podfile_lock.close
+      @bad_yaml_podfile_lock.close
+      @conflict_podfile_lock.close
+    end
 
     describe YAMLHelper do
 
@@ -137,22 +88,42 @@ module Pod
 
     #-------------------------------------------------------------------------#
 
-    describe 'Loading' do
+    describe 'Loading Strings' do
       it 'raises an Informative error when it encounters a merge conflict' do
         should.raise Informative do
-          YAMLHelper.load(yaml_with_merge_conflict)
-        end.message.should.match /Merge conflict\(s\) detected/
+          YAMLHelper.load_string(@conflict_podfile_lock.read)
+        end.message.should.match /unable to continue due to merge conflicts/
       end
 
-      it 'raises another error when it encounters an error that is not a merge conflict' do
+      it 'raises error when encountering a non-merge conflict error' do
         should.raise Exception do
-          YAMLHelper.load(bad_yaml)
+          YAMLHelper.load_string(@bad_yaml_podfile_lock.read)
         end
       end
 
       it 'should not raise when there is no merge conflict' do
         should.not.raise do
-          YAMLHelper.load(sample_yaml)
+          YAMLHelper.load_string(@good_podfile_lock.read)
+        end
+      end
+    end
+
+    describe 'Loading Files' do
+      it 'raises an Informative error when it encounters a merge conflict' do
+        should.raise Informative do
+          YAMLHelper.load_file(Pathname.new(@conflict_podfile_lock.path))
+        end.message.should.match /unable to continue due to merge conflicts/
+      end
+
+      it 'raises error when it encounters a non-merge conflict error' do
+        should.raise Exception do
+          YAMLHelper.load_file(Pathname.new(@bad_yaml_podfile_lock.path))
+        end
+      end
+
+      it 'should not raise when there is no merge conflict' do
+        should.not.raise do
+          YAMLHelper.load_file(Pathname.new(@good_podfile_lock.path))
         end
       end
     end
@@ -229,17 +200,18 @@ module Pod
     #-------------------------------------------------------------------------#
 
     describe 'Lockfile generation' do
-
       it 'converts a complex file' do
-        value = YAMLHelper.load(sample_yaml)
-        sorted_keys = ['PODS', 'DEPENDENCIES', 'EXTERNAL SOURCES', 'SPEC CHECKSUMS', 'COCOAPODS']
+        podfile_str = @good_podfile_lock.read
+        value = YAMLHelper.load_string(podfile_str)
+        sorted_keys = ['PODS', 'DEPENDENCIES', 'SPEC CHECKSUMS', 'COCOAPODS']
         result = YAMLHelper.convert_hash(value, sorted_keys, "\n\n")
-        YAMLHelper.load(result).should == value
-        result.should == sample_yaml
+        YAMLHelper.load_string(result).should == value
+        result.should == podfile_str
       end
     end
 
     #-------------------------------------------------------------------------#
-
   end
+
+  #---------------------------------------------------------------------------#
 end